# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Hash Map API",
        "version": "v1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/get/{key}": {
            "get": {
                "tags": [
                    "HashMap"
                ],
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/set": {
            "post": {
                "tags": [
                    "HashMap"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "key": {
                                        "type": "string"
                                    },
                                    "value": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success"
                    }
                }
            }
        }
    }
}